Add shared interface folder and move Services folder to root project.
