+++ ShelterVault/ShelterVault/Services/EncryptionService.cs
