+++ ShelterVault/ShelterVault/Services/MasterKeyService.cs
