+++ ShelterVault/ShelterVault/Services/ProgressBarService.cs
