+++ ShelterVault/ShelterVault/Shared/Behaviors/EnableNavigationViewModelBehavior.cs
@@ -1,7 +1,7 @@
 ﻿using Microsoft.UI.Xaml;
 using Microsoft.UI.Xaml.Controls;
 using Microsoft.Xaml.Interactivity;
-using ShelterVault.Shared.Services;
+using ShelterVault.Shared.Interfaces;
 using System;
 using System.Collections.Generic;
 using System.Linq;
